refactor: move helper files to helpers/
